backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
gunicorn==21.2.0
google-cloud-bigquery==3.14.1
db-dtypes==1.2.0
pandas==2.1.4
pyarrow>=14.0.0
elasticsearch==8.12.0
requests==2.31.0
